Title: 최솟값 3개, Time: 704ms, Memory: 22MB, Status: Passed - CodeTree

+++ 250226/최솟값 3개/top-3-smallest-number.java
@@ -18,19 +18,19 @@ public class Main {
         StringTokenizer st = new StringTokenizer(br.readLine());
         n = Integer.parseInt(st.nextToken());
 
-        PriorityQueue<Integer> pq = new PriorityQueue<>();
+        PriorityQueue<Long> pq = new PriorityQueue<>();
 
         // 가장 작은 숫자 3개 곱
 
         st = new StringTokenizer(br.readLine());
         for (int i = 0; i < n; i++) {
-            int x = Integer.parseInt(st.nextToken());
+            long x = Integer.parseInt(st.nextToken());
             pq.add(x);
 
             if (pq.size() >= 3){
-                int a = pq.poll();
-                int b = pq.poll();
-                int c = pq.poll();
+                long a = pq.poll();
+                long b = pq.poll();
+                long c = pq.poll();
                 System.out.println(a*b*c);
                 pq.add(a); pq.add(b); pq.add(c);
             }else
